refactor: モジュールを責務別サブパッケージに再編する
21 個のフラットなモジュールを audio / transcription / minutes / core の
4 サブパッケージに分割し、コードベースのナビゲーション性を向上させる。

- audio/    : stream, writer, preprocess, vad, devices
- transcription/ : live, live_transcription, transcribe, filter, dedupe
- minutes/  : summarize, clean, ollama_client, prompts
- core/     : output, metadata, vocabulary, checks
- トップレベルに cli, config, errors を残す

ファイル名も一部整理（audio_output→writer, transcript_filter→filter）。
全 92 テスト通過済み。
