+++ src/meeting_minutes/core/checks.py
@@ -7,7 +7,7 @@ import httpx
 import sounddevice as sd
 
 from meeting_minutes.config import AppConfig
-from meeting_minutes.devices import list_input_devices
+from meeting_minutes.audio.devices import list_input_devices
 
 logger = logging.getLogger(__name__)
