+++ src/meeting_minutes/core/vocabulary.py
