+++ src/meeting_minutes/transcription/live_transcription.py
@@ -4,10 +4,10 @@ from typing import Protocol
 
 import numpy as np
 
-from meeting_minutes.dedupe import TranscriptDedupe
-from meeting_minutes.transcribe import TranscriptionSegment
-from meeting_minutes.transcript_filter import TranscriptFilter
-from meeting_minutes.vad import SpeechSegment, SpeechSegmenter
+from meeting_minutes.transcription.dedupe import TranscriptDedupe
+from meeting_minutes.transcription.transcribe import TranscriptionSegment
+from meeting_minutes.transcription.filter import TranscriptFilter
+from meeting_minutes.audio.vad import SpeechSegment, SpeechSegmenter
 
 
 class SegmentWriter(Protocol):
