+++ tests/test_metadata.py
@@ -5,8 +5,8 @@ from pathlib import Path
 import pytest
 
 from meeting_minutes.config import load_config
-from meeting_minutes.devices import InputDevice
-from meeting_minutes.metadata import build_metadata, write_metadata
+from meeting_minutes.audio.devices import InputDevice
+from meeting_minutes.core.metadata import build_metadata, write_metadata
 
 
 def test_write_metadata_serializes_datetime_and_path(tmp_path: Path) -> None:
