+++ tests/test_preprocess.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from meeting_minutes.config import PreprocessingConfig
-from meeting_minutes.preprocess import AudioPreprocessor
+from meeting_minutes.audio.preprocess import AudioPreprocessor
 
 
 def test_preprocessor_is_noop_when_disabled() -> None:
